Move PointsList componet to admin directory
