+++ app/javascript/packs/application/components/admin/PointsList.test.js
