+++ app/javascript/packs/application/components/admin/TourContainer.js
@@ -4,7 +4,7 @@ import { Link } from "@reach/router";
 import { connect } from "react-redux";
 import Tour from "../Tour";
 import AddPointForm from "./AddPointForm";
-import PointsList from "../PointsList";
+import PointsList from "./PointsList";
 import * as tourActions from "../../redux/tours/actions";
 
 export class TourContainer extends Component {
